Only test sqlite behavior on non-windows platforms
